+++ src/providers/provider.rs
@@ -1,5 +1,6 @@
 use reqwest::Url;
 use async_trait::async_trait;
+use rust_decimal::Decimal;
 
 use crate::currency::Symbol;
 
@@ -12,5 +13,5 @@ pub struct BaseProvider {
 #[async_trait]
 pub(crate) trait Provider {
     fn build_url(&self, base: &Symbol, quote: &Symbol) -> Result<Url, anyhow::Error>;
-    async fn get_rate(&self, base: Symbol, quote: Symbol) -> Result<f64, anyhow::Error>;
+    async fn get_rate(&self, base: Symbol, quote: Symbol) -> Result<Decimal, anyhow::Error>;
 }
